Change naming, rename models to repositories
