+++ internal/repositories/email.go
@@ -1,4 +1,4 @@
-package models
+package repositories
 
 import (
 	"database/sql"
@@ -8,11 +8,11 @@ import (
 	"github.com/lib/pq"
 )
 
-type EmailModelPostgres struct {
+type PostgresEmailRepository struct {
 	DB *sql.DB
 }
 
-func (em *EmailModelPostgres) Insert(email string) error {
+func (em *PostgresEmailRepository) Insert(email string) error {
 	stmt := `INSERT INTO emails (email) VALUES ($1)`
 
 	_, err := em.DB.Exec(stmt, email)
@@ -29,7 +29,7 @@ func (em *EmailModelPostgres) Insert(email string) error {
 	return nil
 }
 
-func (em *EmailModelPostgres) GetAll() ([]string, error) {
+func (em *PostgresEmailRepository) GetAll() ([]string, error) {
 	query := `SELECT email FROM emails`
 	var emails []string
